code to percolate the expression to the GridView is there, just need to call it from GridView.

+++ GraphCalc/GraphView.h
@@ -12,7 +12,7 @@
 
 @protocol GraphViewDelegate
 @property (nonatomic) int scale;
-- (int)YValueForX:(int)x;
+- (double)YValueForX:(int)x;
 @end
